+++ doc/CHANGES.rst
@@ -11,6 +11,9 @@ http://docs.zope.org/zope2/releases/.
 Bugs Fixed
 ++++++++++
 
+- Document ``Products.PluginIndexes.PathIndex.PathIndex.insertEntry`` as
+  an API for use by subclasses.
+
 - Fixed regression introduced in the resource directory fix in 2.12.4.
 
 - LP #143655:  don't prevent sorting using a path index.